add note of my component and delete shouldUpdate and trim the showRoute functions
