+++ package.json
@@ -38,6 +38,7 @@
   },
   "devDependencies": {
     "autoprefixer": "^6.3.3",
+    "axios": "^0.12.0",
     "babel-core": "^6.5.2",
     "babel-eslint": "^5.0.0",
     "babel-loader": "^6.2.3",
@@ -54,6 +55,7 @@
     "classnames": "^2.2.3",
     "clean-webpack-plugin": "^0.1.8",
     "css-loader": "^0.23.1",
+    "es6-promise": "^3.2.1",
     "eslint": "~2.2.0",
     "eslint-plugin-react": "^4.1.0",
     "express": "^4.13.4",
@@ -67,7 +69,7 @@
     "node-sass": "^3.4.2",
     "postcss-loader": "^0.8.1",
     "react-cookie": "^0.4.5",
-    "react-custom-scrollbars":"^4.0.0-beta.2",
+    "react-custom-scrollbars": "^4.0.0-beta.2",
     "react-hot-loader": "^1.3.0",
     "react-router": "^2.0.0",
     "react-transform-catch-errors": "^1.0.2",
