+++ src/UIComponent/Modals/Modal.js
@@ -1,3 +1,10 @@
+/*
+ *  Project : Model
+ *  Date    : 2016.06.28
+ *  Author  : Paco
+ *  Declare : Provide A Model Container
+ *  UseAge  : Need a component to show and className to the model
+ */
 import React, {PropTypes}  from 'react';
 
 import styles from './Modal.scss';
