+++ src/UIComponent/Tree/Tree.js
@@ -1,3 +1,9 @@
+/*
+ *  Project : Tree
+ *  Date    : 2016.06.28
+ *  Author  : Paco
+ *  Declare : Like the Menu, but it can show what you want to show and to choose it.
+ */
 'use strict';
 import React, {PropTypes} from 'react';
 import {connect} from 'react-redux';
